+++ SocialMediaPlatform.Business/Interfaces/IUserService.cs
@@ -6,7 +6,7 @@ public interface IUserService
 {
     Task<UserDto> GetUserProfileAsync(Guid userId, Guid? currentUserId = null);
     Task<UserDto> GetUserByUserNameAsync(string userName, Guid? currentUserId = null);
-    Task<UserDto> UpdateUserProfileAsync(Guid userId, UpdateUserProfileDto updateDto);
+    Task<UserDto> UpdateUserProfileAsync(Guid userId, UserProfileUpdateDto userProfileUpdateDto);
     Task<List<UserSummaryDto>> SearchUsersAsync(string searchTerm, int page = 1, int pageSize = 20);
     Task UpdateLastLoginAsync(Guid userId);
 }
