+++ SocialMediaPlatform.Entities/Dtos/AuthDtos/RegisterRequest.cs
@@ -1,9 +1,9 @@
-namespace SocialMediaPlatform.Entities.Dtos;
+namespace SocialMediaPlatform.Entities.Dtos.AuthDtos;
 
 public class RegisterRequest
 {
     public required string Email { get; set; }
     public required string Username { get; set; }
-    public required string FullName { get; set; }
+    public  string? FullName { get; set; }
     public required string Password { get; set; }
 }
