+++ SocialMediaPlatform.Entities/Dtos/CommentDtos/CommentDto.cs
@@ -1,25 +1,24 @@
-namespace SocialMediaPlatform.Entities.Dtos;
+using SocialMediaPlatform.Entities.Dtos.UserDtos;
+
+namespace SocialMediaPlatform.Entities.Dtos.CommentDtos;
 
 public class CommentDto
 {
-    public int Id { get; set; }
-    public int PostId { get; set; }
+    public Guid Id { get; set; }
+    public Guid PostId { get; set; }
     public Guid AuthorId { get; set; }
-    public int? ParentCommentId { get; set; }
+    public Guid? ParentCommentId { get; set; }
     public string Content { get; set; } = null!;
-    public DateTime CreatedAt { get; set; }
+    public DateTimeOffset CreatedAt { get; set; }
+    public DateTimeOffset? UpdatedAt { get; set; }
     
-    // Author information
     public UserSummaryDto Author { get; set; } = null!;
     
-    // Statistics
     public int LikesCount { get; set; }
     public int RepliesCount { get; set; }
     
-    // Current user's interaction
     public bool IsLikedByCurrentUser { get; set; }
     
-    // Replies (can be included or excluded based on needs)
     public List<CommentDto>? Replies { get; set; }
 }
