--- SocialMediaPlatform.Entities/Dtos/CreateCommentDto.cs
@@ -1,9 +1,0 @@
-namespace SocialMediaPlatform.Entities.Dtos
-{
-    public class CreateCommentDto
-    {
-        public int PostId { get; set; }
-        public required string Content { get; set; }
-        public int? ParentCommentId { get; set; }
-    }
-}
