--- SocialMediaPlatform.Entities/Dtos/CreateStoryDto.cs
@@ -1,8 +1,0 @@
-using Microsoft.AspNetCore.Http;
-
-namespace SocialMediaPlatform.Entities.Dtos.Story;
-
-public class CreateStoryDto
-{
-    public IFormFile MediaFile { get; set; }
-}
