+++ SocialMediaPlatform.Entities/Dtos/FollowDtos/FollowDto.cs
@@ -1,15 +1,15 @@
+using SocialMediaPlatform.Entities.Dtos.UserDtos;
 using SocialMediaPlatform.Entities.Models;
 
-namespace SocialMediaPlatform.Entities.Dtos;
+namespace SocialMediaPlatform.Entities.Dtos.FollowDtos;
 
 public class FollowDto
 {
     public Guid FollowerId { get; set; }
     public Guid FollowedId { get; set; }
-    public DateTime CreatedAt { get; set; }
-    public DateTime? DecidedAt { get; set; }
+    public DateTimeOffset CreatedAt { get; set; }
+    public DateTimeOffset? DecidedAt { get; set; }
     public FollowStatus Status { get; set; }
-    
     public UserSummaryDto Follower { get; set; } = null!;
     public UserSummaryDto Followed { get; set; } = null!;
 }
