+++ SocialMediaPlatform.Entities/Dtos/MessageDtos/MessageCreateDto.cs
@@ -1,0 +1,7 @@
+namespace SocialMediaPlatform.Entities.Dtos.MessageDtos;
+
+public class MessageCreateDto
+{
+    public Guid ReceiverId { get; set; }
+    public string? Content { get; set; }
+}
