+++ SocialMediaPlatform.Entities/Dtos/NotificationDtos/NotificationDto.cs
@@ -1,17 +1,16 @@
-using System;
+using SocialMediaPlatform.Entities.Dtos.UserDtos;
 using SocialMediaPlatform.Entities.Models;
 
-namespace SocialMediaPlatform.Entities.Dtos;
+namespace SocialMediaPlatform.Entities.Dtos.NotificationDtos;
 
 public class NotificationDto
 {
-    public int Id { get; set; }
+    public Guid Id { get; set; }
     public Guid RecipientId { get; set; }
     public NotificationType Type { get; set; }
     public string Message { get; set; } = null!;
-    public string? ActionUrl { get; set; }
     public bool IsRead { get; set; }
-    public DateTime CreatedAt { get; set; }
+    public DateTimeOffset CreatedAt { get; set; }
     
     // Related entities (optional, depending on notification type)
     public UserSummaryDto? Actor { get; set; }  // User who triggered the notification
