+++ SocialMediaPlatform.Entities/Dtos/PostDtos/PostCreateDto.cs
@@ -1,0 +1,9 @@
+using Microsoft.AspNetCore.Http;
+
+namespace SocialMediaPlatform.Entities.Dtos.PostDtos;
+
+public class PostCreateDto
+{
+    public string? Caption { get; set; }
+    public required List<IFormFile> MediaFiles { get; set; }
+}
