+++ SocialMediaPlatform.Entities/Dtos/PostDtos/PostUpdateDto.cs
@@ -1,0 +1,6 @@
+namespace SocialMediaPlatform.Entities.Dtos.PostDtos;
+
+public class PostUpdateDto
+{
+    public string? Caption { get; set; }
+}
