+++ SocialMediaPlatform.Entities/Dtos/StoryDtos/StoryCreateDto.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Http;
+
+namespace SocialMediaPlatform.Entities.Dtos.StoryDtos;
+
+public class StoryCreateDto
+{
+    public required IFormFile MediaFile { get; set; }
+}
