+++ SocialMediaPlatform.Entities/Dtos/UserDtos/UserSummaryDto.cs
@@ -1,0 +1,9 @@
+namespace SocialMediaPlatform.Entities.Dtos.UserDtos;
+
+public class UserSummaryDto
+{
+    public Guid Id { get; set; }
+    public string UserName { get; set; } = null!;
+    public string? FullName { get; set; }
+    public string? ProfilePictureUrl { get; set; }
+}
